- Feat
Slider Function for Earring & Ring
Logic for the buff/debuff.
Enemy AI Stop Moving/Roaming around near player when player active the jewels. if you want to make it so the enemy stop near player or still patrolling then need to adjust the AI for walkPoint.

WIP : Some of the function were not ready yet. cause i didnt find the assets. like audio. and as for the blur effect sometimes it works and sometimes it's not. need to check it

Door interaction.

ps : for now i destroy the door. but if you want to add an animation for it i can do it as long as the animation is simple.